requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asyncpg==0.30.0
attrs==24.3.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
fake-useragent==2.0.3
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.2.1
pandas==2.2.3
propcache==0.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
tzdata==2024.2
urllib3==2.3.0
yarl==1.18.3
